+++ js/main.js
@@ -1,6 +1,3 @@
-import { thumbnailRender } from './thumbnail';
-import { initThumbnailClicks } from './imageViewer';
-thumbnailRender();
-initThumbnailClicks();
+
